+++ src/GeneratorDanych/TavernFactory.java
@@ -1,10 +1,12 @@
 package GeneratorDanych;
 
+import SiecPrzeplywowa.Tavern;
+
 import java.util.Random;
 
 class TavernFactory implements IFactory<Tavern> {
     @Override
-    public Tavern Create(Random random, double x, double y) {
+    public Tavern Create(Random random, int x, int y) {
         return new Tavern(x, y);
     }
 }
